requirements.txt
streamlit>=1.20
openai>=0.28.0
python-docx>=0.8.11
pdfplumber>=0.7.6
pandas>=1.5
numpy>=1.23
scikit-learn>=1.2
nltk>=3.8
python-magic>=0.4.27
tqdm>=4.65.0
